Move watch util to client package.
